+++ src/middlewares/http.logger.ts
@@ -1,5 +1,5 @@
 import { Request, Response, NextFunction } from "express";
-import { logger } from "../../utils/logger";
+import { logger } from "../utils/logger";
 
 export const httpLogger = (req: Request, res: Response, next: NextFunction) => {
   logger.info("\ue66a Incoming Request", {