+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## 0.1
+- Add category CRUD
+- Flesh out account CRUD
+- Improve logging
 - Add logging utility
 - Add folder structure
 - Add initial configs and setup